pyproject.toml
[tool.poetry]
name = "apply-test"
version = "0.1.0"
description = ""
authors = ["zwickvitaly <zwickvitaly@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
aiohttp = "^3.9.5"
pillow = "^10.4.0"
loguru = "^0.7.2"
uvicorn = "^0.30.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
